Update proxmox.py to for new numbering scheme
With proxmox 6 they changed the way that the numbers get returned from the API.  There is now a hyphen in it.  This is the same application as in https://github.com/ansible/ansible/pull/59356 but changed to a hypnen.  The other patch is only good for KVM, this fixes it for LXC.